Calculate user achievements and stats locally; achievements init
Speculatively updates the user's achievements and stats using code shared with
the Cloud Functions. The server remains the source of truth--it will push the
updated user document to the client once the hooks have run--, but this lets
us update the UI immediately.

This commit also includes fixes to achievements handling on the server side. It
now initializes achievements correctly.