Preprocessing script for Firebase Hosting
firebase deploy will now run the `render-templates.js` preprocessing script on
the documents in `firebase/project/web/public`. Files containing `.template` in
the extension will be rendered using Handlebars with a config map as context.
The results will be saved to the same path, but with `.template` removed. The
purpose is mainly to allow the config options to be locally overriden.