Improvements to `usePlogs`, added `usePaginatedPlogs`
`usePlogs` now dispatches an action that will allow plog-middleware to
unsubscribe from further updates to that plog, if appropriate. (The logic for
doing that is currently unimplemented. It will need to count references, and
there's no obvious harm in leaving it out.)

`usePaginatedPlogs` loads plog data in increments. This will allow us to
simplify some of the logic in plog-middleware, which is currently tracking too
much state.